Database. Замена типа id с int на long

+++ library/src/ru/qixi/api/database/SampleData.java
@@ -2,42 +2,42 @@ package ru.qixi.api.database;
 
 public class SampleData implements IDBEntity {
 
-	private int		_id;
-	private String	mText;
+    private long   _id;
+    private String mText;
 
 
-	public SampleData() {}
+    public SampleData() {}
 
 
-	public SampleData(String text) {
-		mText = text;
-	}
+    public SampleData(String text) {
+        mText = text;
+    }
 
 
-	public SampleRepository getRepository() {
-		return SampleRepository.getInstance();
-	}
+    public SampleRepository getRepository() {
+        return SampleRepository.getInstance();
+    }
 
 
-	public int getId() {
-		return _id;
-	}
+    public long getId() {
+        return _id;
+    }
 
 
-	public String getText() {
-		return mText;
-	}
+    public String getText() {
+        return mText;
+    }
 
 
-	SampleData setId(int id) {
-		_id = id;
-		return this;
-	}
+    SampleData setId(long id) {
+        _id = id;
+        return this;
+    }
 
 
-	public SampleData setText(String text) {
-		mText = text;
-		return this;
-	}
+    public SampleData setText(String text) {
+        mText = text;
+        return this;
+    }
 
 }
